Revert "adding common cli flag check routines to plugin lib"
This reverts commit eb19e0627f7c98000c91b9002821ff4f4d89ad4d.

We can't put `cli` types in the public interface of a shared package.
This is because (for example) pluginlib's `cli.Context` is
actually a different type than omg-product-bundle's `cli.Context`.

For now (while all product plugins are in a single repo), we have a
`cli` package in omg-product-bundle.  This is the proper location
for such utility functions.  When we break plugins out into spearate
repos we will either have to duplicate these routines in each plugin
repo or abandon vendoring and build using pluginlib from $GOPATH.