Skip plugin registry tests when -short flag is used
These tests are meant to be run from the testrunner script, which sets up pre-requisites for the tests and performs the appropriate clean up.

If the `-short` flag is passed to `go test`, we skip these tests.  This allows us to run the majority of the tests from the command line (or Atom) without running into failures.